requirements.txt
django
djangorestframework
